Move the kiosk mode out of the volunteer module
Also move more of the inner working into its own set of methods.
This will make it possible to add kiosk functionality to other
areas of the code. The current vision has kisok functionality
in a registration module as well.
