+++ modules/volunteers/kiosk/pages/head.inc
@@ -1,7 +1,2 @@
-<script src="sitesupport/barcode/quagga.min.js"></script>
-<script src="sitesupport/barcode/barcode.js"></script>
-<script src="sitesupport/common.js"></script>
-<script src="modules/volunteers/sitesupport/volunteers.js"></script>
-<script src="sitesupport/authentication.js"></script>
-<script src="sitesupport/lookupuser.js"></script>
-<link rel="stylesheet" type="text/css" href="sitesupport/barcode/barcode.css" />
+<?php
+include_once $BASEDIR.'/kiosk/head.inc';
