+++ modules/volunteers/pages/menubar.inc
@@ -1,4 +1,7 @@
 <?php
+
+require_once $BASEDIR.'/kiosk/kiosk.inc';
+
 if (isset($_SESSION['customFields']['currConComPos'])) {
     $base_menus[] = array('base style' => '',
      'selected style' => 'event-color-primary',
@@ -9,14 +12,24 @@ if (isset($_SESSION['customFields']['currConComPos'])) {
      'responsive' => true);
 }
 
-if ($_SESSION['IS_VOLUNTEERS'] || !empty($_SESSION['IS_ADMIN']) ||
-    isset($_SESSION['customFields']['currConComPos'])) {
-    $basecolor = 'w3-yellow';
-    if (array_key_exists('CIAB_VOLUNTEERADMIN', $_COOKIE) &&
-       $_COOKIE["CIAB_VOLUNTEERADMIN"]) {
-        $basecolor = 'w3-red';
+if (!is_kiosk('volunteers')) {
+    if ($_SESSION['IS_VOLUNTEERS'] || !empty($_SESSION['IS_ADMIN']) ||
+        isset($_SESSION['customFields']['currConComPos'])) {
+        $basecolor = 'w3-yellow';
+        if (array_key_exists('CIAB_VOLUNTEERADMIN', $_COOKIE) &&
+           $_COOKIE["CIAB_VOLUNTEERADMIN"]) {
+            $basecolor = 'w3-red';
+        }
+        $admin_menus[] = array('base style' => $basecolor,
+         'selected style' => 'event-color-primary',
+         'function' => 'volunteers/admin',
+         'title' => 'Volunteer\'s',
+         'i class' => 'fa fa-users',
+         'text' => 'Volunteers',
+         'responsive' => true);
     }
-    $admin_menus[] = array('base style' => $basecolor,
+} else {
+    $base_menus[] = array('base style' => '',
      'selected style' => 'event-color-primary',
      'function' => 'volunteers/admin',
      'title' => 'Volunteer\'s',
@@ -25,12 +38,4 @@ if ($_SESSION['IS_VOLUNTEERS'] || !empty($_SESSION['IS_ADMIN']) ||
      'responsive' => true);
 }
 
-if (isset($_SESSION['kioskMode'])) {
-    $admin_menus[] = array('base style' => 'w3-right event-secondary-color',
-     'selected style' => 'event-color-primary w3-right',
-     'function' => 'volunteers/kiosk',
-     'title' => 'Change Kiosk User from '.$_SESSION['kioskUser'],
-     'i class' => 'fa fa-exchange',
-     'text' => 'Change Kiosk User from '.$_SESSION['kioskUser'],
-     'responsive' => true);
-}
+generate_kiosk_menus('volunteers');
