+++ modules/volunteers/sitesupport/volunteers.js
@@ -5,7 +5,7 @@
 /* jshint browser: true */
 /* jshint -W097 */
 /* globals confirmbox, userId, escapeHtml, userEmail, checkAuthentication,
-           kioskMode, groupData, checkAuthentication, adminMode, unclaimed,
+           groupData, checkAuthentication, adminMode, unclaimed,
            hoursRemain, userLookup */
 
 'use strict';
@@ -560,36 +560,6 @@ function departmentReport(name, dept) {
   document.getElementById('dept_name').innerHTML = name;
   document.getElementById('dept_data').value = dept;
   document.getElementById('dept_data_name').value = name;
-}
-
-function switchKiosk() {
-  var xhttp = new XMLHttpRequest();
-  xhttp.onreadystatechange = function() {
-    if (this.readyState == 4 && this.status == 200) {
-      setTimeout(function() {location.reload() ;}, 1000);
-    }
-  };
-  xhttp.open('POST', 'index.php?Function=volunteers/kiosk', true);
-  xhttp.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-  xhttp.send('&toggleKiosk=true');
-}
-
-function failKiosk(error) {
-  document.getElementById('kiosk_slider').checked = true;
-  if (error) {
-    window.alert('Login Failed (' + error + ')');
-  }
-}
-
-function toggleKioskMode() {
-  if (kioskMode) {
-    checkAuthentication(userEmail, switchKiosk, failKiosk,
-                        {title: 'Exiting Kiosk Mode'});
-  } else {
-    setTimeout(function() {
-        window.location = 'index.php?Function=volunteers/kiosk';
-      }, 1000);
-  }
 }
 
 function generateCSV() {
